requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.13.0
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
cloudscraper==1.2.71
cryptography==46.0.7
distro==1.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-genai==1.72.0
google-generativeai==0.8.6
googleapis-common-protos==1.74.0
greenlet==3.4.0
grpcio==1.80.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
inquirerpy==0.3.4
packaging==26.0
pfzy==0.3.4
playwright==1.58.0
prompt_toolkit==3.0.52
proto-plus==1.27.2
protobuf==5.29.6
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
pyparsing==3.3.2
python-dotenv==1.2.2
requests==2.33.1
requests-toolbelt==1.0.0
setuptools==82.0.1
sniffio==1.3.1
soupsieve==2.8.3
tenacity==9.1.4
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.6.3
wcwidth==0.6.0
websockets==16.0
yt-dlp==2026.3.17
